feat(validator, footer): implement tooling bar
