+++ core/common/src/commonMain/kotlin/com/neoutils/neoregex/core/common/extension/TextState.kt
@@ -16,18 +16,17 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.neoutils.neoregex.core.sharedui.extension
+package com.neoutils.neoregex.core.common.extension
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.input.TextFieldValue
+import com.neoutils.neoregex.core.common.model.Text
 
-fun String.withSpanStyles(
-    spanStyles: List<AnnotatedString.Range<SpanStyle>>
-): AnnotatedString {
-    return AnnotatedString(
-        text = this,
-        spanStyles = spanStyles.filter {
-            it.start <= length && it.end <= length
-        }
-    )
-}
+fun Text.toTextFieldValue(
+    spanStyles: List<AnnotatedString.Range<SpanStyle>> = listOf()
+) = TextFieldValue(
+    annotatedString = text.withSpanStyles(spanStyles),
+    selection = selection,
+    composition = null
+)
