+++ core/shared-ui/src/androidMain/kotlin/com/neoutils/neoregex/core/sharedui/component/TextEditor.android.kt
@@ -43,13 +43,12 @@ import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.text.TextLayoutResult
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.LineHeightStyle
-import com.neoutils.neoregex.core.common.extension.toTextState
+import com.neoutils.neoregex.core.common.extension.toText
 import com.neoutils.neoregex.core.common.model.Text
 import com.neoutils.neoregex.core.designsystem.theme.NeoTheme.dimensions
 import com.neoutils.neoregex.core.sharedui.extension.getBoundingBoxes
-import com.neoutils.neoregex.core.sharedui.extension.toTextFieldValue
+import com.neoutils.neoregex.core.common.extension.toTextFieldValue
 import com.neoutils.neoregex.core.sharedui.model.Match
 import com.neoutils.neoregex.core.sharedui.model.MatchBox
 
@@ -138,7 +137,7 @@ actual fun TextEditor(
         BasicTextField(
             value = textFileValue,
             onValueChange  = {
-                onValueChange(it.toTextState())
+                onValueChange(it.toText())
             },
             textStyle = mergedTextStyle.copy(
                 lineHeightStyle = LineHeightStyle(
